Finished moving from scope to state
