+++ dvelop-sdk-logging/Logging.Abstractions/State/IncomingHttpRequestLogState.cs
@@ -1,15 +1,15 @@
 ﻿using System.Collections.Generic;
-using Dvelop.Sdk.Logging.Abstractions.State;
 using Dvelop.Sdk.Logging.Abstractions.State.Attribute;
 
-namespace Dvelop.Sdk.Logging.Abstractions.Scope
+namespace Dvelop.Sdk.Logging.Abstractions.State
 {
-    public class OutgoingHttpRequestLogState : CustomLogAttributeState
+    public class IncomingHttpRequestLogState : CustomLogAttributeState
     {
         public string Method { get; set; }
         public string Target { get; set; }
-        public int? StatusCode { get; set; }
-        public long? ClientDuration { get; set; }
+        public string UserAgent { get; set; }
+        public int? Status { get; set; }
+        public long? ServerDuration { get; set; }
 
         public override IEnumerable<CustomLogAttribute> Attributes
         {
@@ -26,25 +26,28 @@ namespace Dvelop.Sdk.Logging.Abstractions.Scope
                 {
                     httpAttributes.Add(new CustomLogAttributeProperty("target", Target));
                 }
-
-                if (StatusCode.HasValue)
+                
+                if (!string.IsNullOrEmpty(UserAgent))
                 {
-                    httpAttributes.Add(new CustomLogAttributeProperty("statusCode", StatusCode.Value));
+                    httpAttributes.Add(new CustomLogAttributeProperty("userAgent", UserAgent));
                 }
 
-                if (ClientDuration.HasValue)
+                if (Status.HasValue)
                 {
-                    httpAttributes.Add(new CustomLogAttributeObject("client", new List<CustomLogAttribute>{
-                        new CustomLogAttributeProperty("duration", ClientDuration.Value)
+                    httpAttributes.Add(new CustomLogAttributeProperty("status", Status.Value));
+                }
+
+                if (ServerDuration.HasValue)
+                {
+                    httpAttributes.Add(new CustomLogAttributeObject("server", new List<CustomLogAttribute>{
+                        new CustomLogAttributeProperty("duration", ServerDuration.Value)
                     }));
                 }
 
-                httpAttributes.Add(new CustomLogAttributeProperty("direction", "outbound"));
+                httpAttributes.Add(new CustomLogAttributeProperty("direction", "inbound"));
                 attributes.Add(new CustomLogAttributeObject("http", httpAttributes));
                 return attributes;
             }
         }
-
-        
     }
 }
